requirements.txt
Flask
google-api-python-client
gspread
oauth2client
Pillow
python-dotenv
smtplib
